Updates `IronBank::Client#connection`
Updates `IronBank::Client#connection` to specify internal middlewares by class
name, instead of registering them with Faraday. This fixes a naming conflict
with Faraday's `Faraday::Response::RaiseError` middleware, which was also
registered using the `:raise_error` name.
